requirements.txt
# AutoGen - Multi-Agent Orchestration
pyautogen>=0.2.0

# LangGraph - StateGraph + Memory Management
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# Google Gemini LLM
google-generativeai>=0.8.0

# Web Search Tool (free, no API key)
duckduckgo-search>=6.0.0

# UI
streamlit>=1.35.0
plotly>=5.18.0

# Data & Utilities
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fpdf2>=2.7.0
